refactor: move sandbox js code to its own module
